fix(build): signedness in cdx_index seek_key binary search
The CDX B-tree descent (1952d5c2) indexed a branch entry with
`mid * stride` (int * size_t); clang -Werror,-Wsign-conversion failed
every POSIX CI job. MSVC does not warn, so windows-x64/x86 stayed
green. Cast mid to size_t, matching the sibling child-pointer calc.

+++ src/drivers/cdx/cdx_index.cpp
@@ -837,7 +837,8 @@ CdxIndex::seek_key(const std::string& key, bool soft) {
         int lo = 0, hi = nkeys - 1;
         while (lo < hi) {
             int mid = (lo + hi) >> 1;
-            const std::uint8_t* ek = base + CDX_INT_HEADSIZE + mid * stride;
+            const std::uint8_t* ek = base + CDX_INT_HEADSIZE +
+                static_cast<std::size_t>(mid) * stride;
             int cmp = compare_keys_(padded.data(), reinterpret_cast<const char*>(ek), cmp_len);
             if (cmp > 0)
                 lo = mid + 1;
